+++ bin/adhan.dart
@@ -8,9 +8,9 @@ void main() {
   print(
       'Adhan for Dart / Muslim Prayer Times Library. Now retrieving Prayer time in Dart easier than ever.');
   print('What is your Latitude?');
-  var lat = stdin.readLineSync(encoding: Encoding.getByName('utf-8'));
+  var lat = stdin.readLineSync(encoding: Encoding.getByName('utf-8')!)!;
   print('What is your Longitude?');
-  var lng = stdin.readLineSync(encoding: Encoding.getByName('utf-8'));
+  var lng = stdin.readLineSync(encoding: Encoding.getByName('utf-8')!)!;
 
   final coordinates = Coordinates(double.parse(lat), double.parse(lng));
   final params = CalculationMethod.karachi.getParameters();